+++ sites/mytril.dev/src/lib/store/router/state.ts
@@ -3,7 +3,7 @@ import { writable } from 'svelte/store';
 export const navigation = writable([
 	{
 		name: 'docs',
-		path: '/docs'
+		path: '/docs/introduction'
 	}
 ]);
 
@@ -11,6 +11,10 @@ export const packageLink = writable([
 	{
 		name: 'README',
 		path: 'https://github.com/minedelve/mytril/blob/main/README.md'
+	},
+	{
+		name: 'CHANGELOG',
+		path: 'https://github.com/minedelve/mytril/blob/main/CHANGELOG.md'
 	}
 ]);
 
@@ -36,7 +40,32 @@ export const docs = writable([
 			{
 				name: 'introduction',
 				path: '/docs/introduction'
+			},
+			{
+				name: 'installation',
+				path: '/docs/installation'
 			}
+		]
+	},
+	{
+		section: 'Customization',
+		routes: [
+			{
+				name: 'themes',
+				path: '/docs/themes'
+			},
+			{
+				name: 'theme switcher',
+				path: '/docs/theme-switcher'
+			},
+			{
+				name: 'colors',
+				path: '/docs/colors'
+			}
+			// {
+			// 	name: 'breakpoints',
+			// 	path: '/docs/breakpoints'
+			// }
 		]
 	},
 	{