+++ sites/mytril.dev/src/lib/utils/index.ts
@@ -1,4 +1,4 @@
-export * from './capitalize';
 export * from './scroll';
 export * from './click';
 export * from './cookie';
+export * from './copy';